complete ranking flow

+++ BigData-AE/src/uk/ac/gla/dcs/bigdata/studentfunctions/DocTermsReducer.java
@@ -13,17 +13,17 @@ public class DocTermsReducer implements ReduceFunction<CleanedArticle> {
 	@Override
 	public CleanedArticle call(CleanedArticle a1, CleanedArticle a2) throws Exception {
 	
-		Map<String,Integer> terms1 = a1.getTerms();
-		Map<String,Integer> terms2 = a2.getTerms();
+		Map<String,Short> terms1 = a1.getTerms();
+		Map<String,Short> terms2 = a2.getTerms();
 		
-		Map<String,Integer> terms3 = new HashMap<>();
+		Map<String,Short> terms3 = new HashMap<>();
 		
 		terms3.putAll(terms1);
 		
 		for(String term: terms2.keySet()) {
-			Integer value = terms2.get(term);
+			Short value = terms2.get(term);
 			if ( terms3.containsKey(term) ) {
-				terms3.put(term, terms3.get(term) + value);
+				terms3.put(term, (short) (terms3.get(term) + value));
 			}
 			else {
 				terms3.put(term, value);
